Created unit test to assert whether JIT mode can handle U and V Fields with different sizes
Since the current implementation of JIT mode only supports one set of
(xi, yi, zi), JIT mode doesn’t work with Fields with different grids.
This is most obvious where the sizes of the grids are widely different,
when an xi on one grid falls outside the domain on another one This
unit test checks for this. A possible solution would be to store
separate xi, yi etc for each Field

Note that it does work in Scipy mode, since interpolation there does
not depend on xi, yi, etc.