Integration of the Windows gcc compiler chain into Conda.
The package `m2w64-toolchain` was found on the Conda repository and it can replace the functionality of `gcc` completely for Parcels with either a 32 or 64-bit Python Conda distribution. This meant that the need for the user to install `mingw-w64` himself was alleviated. To reflect this, changes were also made to `docs/index.rst` to properly show this development.