Adding new script convert_npydir_to_netcdf to create netcdf during run and/or after incomplete run
On every ParticleSet.write(), we now also dump a small pset_info.npy
file in the tempwritedir directory that contains all the relevant
information to construct a netcdf file from the npy files in the
tempwritedir directory.
This means users can 1) create a netcdf file while their run is
executing and 2) create a netcdf file from a run that has not properly
completed (i.e. ParticleSet.__del__ was not called).
This is also possible from the command line