Change method of flattening arrays for JIT cdata
Previously, I was discarding the type information returned from
numpy's ctypeslib. However, it can't properly convert 2D
arrays. Instead, we make sure we get a non-copied flattening view of
the array beforehand, which means that we'll get a regular 1D array
pointer out of ctypeslib. All in all, we shouldn't be passing array
copies to the C code, which would then be discarded after kernel
execution.